dynamo

+++ app/models/movies.py
@@ -1,7 +1,8 @@
+from decimal import Decimal
 from pydantic import BaseModel
 
 
 class Movie(BaseModel):
     id: int
     name: str
-    rating: float
+    rating: Decimal
